refactor : 폴더 생성 후 파일 이동
