+++ src/components/mockInterview/Result/InterviewResult.jsx
