+++ src/components/mockInterview/index.js
@@ -1,6 +1,6 @@
 // Layout 컴포넌트들
-export { default as LoadingScreen } from './Layout/LoadingScreen';
-export { default as ProgressBar } from './Layout/ProgressBar';
+export { default as LoadingScreen } from './mockInterview/Layout/LoadingScreen';
+export { default as ProgressBar } from './mockInterview/Layout/ProgressBar';
 
 // Timer 컴포넌트들
 export { default as CircularTimer } from './Timer/CircularTimer';
@@ -12,7 +12,7 @@ export { default as QuestionCard } from './Question/QuestionCard';
 export { default as VideoPlayer } from './Video/VideoPlayer';
 
 // Audio 컴포넌트들
-export { default as AudioVisualizer } from './Audio/AudioVisualizer';
+export { default as AudioVisualizer } from './mockInterview/Audio/AudioVisualizer';
 
 // Result 컴포넌트들
 export { default as InterviewResult } from './Result/InterviewResult';
