+++ src/styles/mockInterview/globals.css
