+++ src/utils/mockInterview/constants.js
